Safer std::async based dispatch
