+++ CMakeLists.txt
@@ -24,7 +24,6 @@ endif (QBCAPPOW GREATER 31)
 add_library (qrack STATIC
     src/common/parallel_for.cpp
     src/common/rdrandwrapper.cpp
-    src/common/dispatchqueue.cpp
     src/qinterface/arithmetic.cpp
     src/qinterface/gates.cpp
     src/qinterface/logic.cpp
@@ -33,7 +32,6 @@ add_library (qrack STATIC
     src/qinterface/rotational.cpp
     src/qengine/qengine.cpp
     src/qengine/arithmetic.cpp
-    src/qengine/gates.cpp
     src/qengine/state.cpp
     src/qengine/utility.cpp
     src/bitbuffer.cpp
@@ -106,7 +104,6 @@ include ("cmake/Complex8.cmake")
 include ("cmake/Complex_x2.cmake")
 include ("cmake/Pure32.cmake")
 include ("cmake/Boost.cmake")
-include ("cmake/QUnit_CPU_Parallel.cmake")
 include ("cmake/VM6502Q.cmake")
 
 target_compile_definitions (qrack PUBLIC QBCAPPOW=${QBCAPPOW})
@@ -124,7 +121,6 @@ else (QBCAPPOW EQUAL 7)
 endif (QBCAPPOW EQUAL 7)
 message ("Single accuracy is: ${ENABLE_COMPLEX8}")
 message ("Complex_x2/AVX support is: ${ENABLE_COMPLEX_X2}")
-message ("Parallel QUnit over CPU support is: ${ENABLE_QUNIT_CPU_PARALLEL}")
 message ("VM6502Q disassembler support is: ${ENABLE_VM6502Q_DEBUG}")
 
 if (ENABLE_UINT128 AND ENABLE_PURE32)
@@ -183,7 +179,6 @@ install (FILES
     include/common/oclengine.hpp
     include/common/parallel_for.hpp
     include/common/rdrandwrapper.hpp
-    include/common/dispatchqueue.hpp
     DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/qrack/common
     )
